requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
idna==2.9
Pillow==7.1.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
vk==2.0.2
vk-api==11.8.0
pony==0.7.13